Moved GUI-related code to Ui, refactored some variables.
