+++ src/main/java/duke/task/PriorityList.java
@@ -2,9 +2,8 @@ package duke.task;
 
 import duke.command.AddCommand;
 import duke.enums.Numbers;
-import duke.task.TaskList;
+
 import java.util.ArrayList;
-import java.util.Scanner;
 import javafx.util.Pair;
 
 //@@author Dou-Maokang
@@ -14,7 +13,7 @@ import javafx.util.Pair;
 public class PriorityList {
 
     private ArrayList<Integer> priorityList;
-    private int defultPriority = Numbers.FIVE.value;
+    private int defaultPriority = Numbers.FIVE.value;
 
     /**
      * Creates an empty priority list using an array list.
@@ -55,7 +54,7 @@ public class PriorityList {
      */
     public PriorityList addDefaultPriority(duke.command.Command cmd) {
         if (cmd instanceof AddCommand) {
-            priorityList.add(defultPriority);
+            priorityList.add(defaultPriority);
         }
         return new PriorityList(priorityList);
     }
@@ -69,7 +68,7 @@ public class PriorityList {
     public PriorityList addMultiDefaultPriority(int numOfTimes) {
 
         for (int i = Numbers.ZERO.value; i < numOfTimes; i++) {
-            priorityList.add(defultPriority);
+            priorityList.add(defaultPriority);
         }
 
         return new PriorityList(priorityList);
