+++ src/main/java/duke/ui/AddNotesWindow.java
@@ -1,5 +1,6 @@
-package duke;
+package duke.ui;
 
+import duke.ui.MainWindow;
 import javafx.fxml.FXML;
 import javafx.scene.control.TextField;
 import javafx.scene.control.Button;
